refactor group day trips diagnostics api
