+++ examples/quickstart-fr.py
@@ -47,7 +47,7 @@ population_trips = mobility.PopulationGroupDayTrips(
 weekday_plan_steps = population_trips.get()["weekday_plan_steps"].collect()
 
 # You can compute global metrics for weekday trips
-global_metrics = population_trips.weekday_run.evaluate("global_metrics")
+global_metrics = population_trips.weekday_run.results().metrics.aggregate()
 
 # You can plot weekday OD flows, with labels for prominent cities
 weekday_results = population_trips.weekday_run.results()
