+++ mobility/trips/group_day_trips/core/__init__.py
@@ -1,13 +1,19 @@
+from .diagnostics import RunDiagnostics
 from .parameters import Parameters
 from .group_day_trips import PopulationGroupDayTrips
+from .metrics import RunMetrics
 from .results import RunResults
 from .run import Run
 from .run_state import RunState
+from .transitions import RunTransitions
 
 __all__ = [
     "PopulationGroupDayTrips",
     "Parameters",
+    "RunDiagnostics",
+    "RunMetrics",
     "Run",
     "RunResults",
     "RunState",
+    "RunTransitions",
 ]
