+++ mobility/trips/group_day_trips/evaluation/__init__.py
@@ -3,10 +3,9 @@ from .calibration_plan_steps import (
     ObservedCalibrationPlanSteps,
     PopulationWeightedCalibrationPlanSteps,
 )
+from .iteration_metrics import IterationMetricsBuilder, IterationMetricsHistory
 from .model_loss import ModelLoss
 from .model_entropy import ModelEntropy
-from .model_entropy_history import ModelEntropyHistory
-from .model_loss_history import ModelLossHistory
 from .observed_plan_steps import ObservedPlanSteps
 from .public_transport_network_evaluation import PublicTransportNetworkEvaluation
 from .population_weighted_plan_steps import PopulationWeightedPlanSteps
@@ -19,10 +18,10 @@ from .trip_pattern_distribution import (
 
 __all__ = [
     "CarTrafficEvaluation",
+    "IterationMetricsBuilder",
+    "IterationMetricsHistory",
     "ModelEntropy",
-    "ModelEntropyHistory",
     "ModelLoss",
-    "ModelLossHistory",
     "ObservedPlanSteps",
     "ObservedTripPatternDistribution",
     "ObservedCalibrationPlanSteps",
