+++ tests/back/integration/test_010_group_day_trips_results_are_reproducible.py
@@ -42,7 +42,7 @@ def test_010_group_day_trips_results_are_reproducible(test_data):
         )
     )
 
-    metrics_run_1 = pop_trips.weekday_run.evaluate("global_metrics")
+    metrics_run_1 = pop_trips.weekday_run.results().metrics.aggregate()
     
     # Remove the results then re run the model with the same inputs
     pop_trips.remove()
@@ -64,7 +64,7 @@ def test_010_group_day_trips_results_are_reproducible(test_data):
         )
     )
 
-    metrics_run_2 = pop_trips.weekday_run.evaluate("global_metrics")
+    metrics_run_2 = pop_trips.weekday_run.results().metrics.aggregate()
     
     # Compare results between runs
     comparison = (
